Fix right-click issue
The previous iteration of the code would not allow the user to open the
article with a right-click because the href was set as '#'. This fixes
that issue, but it also seems that it's not possible to track when the
user opens the web page in a new tab using a right-click.